long long => int64_t

+++ cuda_functions/nms_3D/src/cuda/nms_kernel.h
@@ -9,7 +9,7 @@ extern "C" {
 int const threadsPerBlock = sizeof(unsigned long long) * 8;
 
 void _nms(int boxes_num, float * boxes_dev,
-          unsigned long long * mask_dev, float nms_overlap_thresh);
+          int64_t * mask_dev, float nms_overlap_thresh);
 
 #ifdef __cplusplus
 }
